`IndexedDbApi` inside `IndexedDbVersion`
